+++ packages/adminui-pub/locale/lang-en.js
@@ -509,6 +509,7 @@ UB.i18nExtend({
   'showFilterPanel': 'Filter',
   'mi_dateFrom': 'Date from',
   'mi_dateTo': 'Date to',
+  'mi_modifyDate': 'Last modified',
   'countST': 'Record count',
   'sumST': 'Sum',
   'maxST': 'Max value',