+++ packages/adminui-pub/locale/lang-id.js
@@ -508,6 +508,7 @@ UB.i18nExtend({
   'showFilterPanel': 'Filter',
   'mi_dateFrom': 'Date from',
   'mi_dateTo': 'Date to',
+  'mi_modifyDate': 'Last modified',
   'countST': 'Record count',
   'sumST': 'Sum',
   'maxST': 'Max value',